+++ demo/src/main/java/com/rockheart/demo/DemoApplication.java
@@ -1,0 +1,13 @@
+package com.rockheart.demo;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class DemoApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(DemoApplication.class, args);
+    }
+
+}
