+++ demo/src/main/java/com/rockheart/demo/service/EmpService.java
@@ -1,0 +1,10 @@
+package com.rockheart.demo.service;
+
+public class EmpService {
+
+    public void add() {
+        //todo
+    }
+
+
+}
