+++ demo/src/main/resources/config/application-dev.properties
@@ -1,0 +1,19 @@
+server.port=8080
+
+#日志级别
+logging.level.com.rockheart=warn
+
+#root日志级别
+logging.level.root=debug
+
+#日志文件路径
+logging.path=springboot/log
+
+logging.pattern.console=%d{yyyy-MM-dd} [%thread] %-5level %logger{50} - %msg%n
+
+logging.pattern.file=%d{yyyy-MM-dd HH:mm:ss.SSS} >>> [%thread] >>> %-5level >>> %logger{50} >>> %msg%n
+
+spring.thymeleaf.prefix=classpath:/template/
+spring.thymeleaf.encoding=utf-8
+spring.thymeleaf.suffix=.html
+spring.thymeleaf.cache=false
