+++ demo/src/main/resources/config/application.yml
@@ -1,0 +1,15 @@
+server:
+  port: 8080 # 默认端口号
+  servlet:
+    context-path: /
+
+spring:
+  profiles:
+    active: dev #激活哪个profile , 当前激活的是 dev 开发环境
+
+  datasource:
+    driver-class-name: com.mysql.jdbc.Driver
+    url:
+
+
+
