--- src/main/java/com/rockheart/GetBootSkillApp.java
@@ -1,14 +1,0 @@
-package com.rockheart;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-//@ImportResource(locations = {"classpath:/config/spring.xml"})
-@SpringBootApplication
-public class GetBootSkillApp {
-
-    public static void main(String[] args) {
-        SpringApplication.run(GetBootSkillApp.class, args);
-    }
-
-}
