--- src/main/java/com/rockheart/po/Emp.java
@@ -1,133 +1,0 @@
-package com.rockheart.po;
-
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.context.annotation.PropertySource;
-import org.springframework.stereotype.Component;
-import org.springframework.validation.annotation.Validated;
-
-import javax.validation.constraints.Email;
-import java.util.Date;
-import java.util.List;
-import java.util.Map;
-
-@Component
-@PropertySource(value={"classpath:/config/emp.yml"})
-@ConfigurationProperties(prefix = "emp")
-@Validated
-public class Emp {
-
-//    @Value("${emp.last-name}")
-    @Email
-    private String lastName;
-//    @Value("#{10*2}")
-    private Integer age;
-//    @Value("8000")
-    private Double salary;
-    private Boolean boss;
-    private Date birthday;
-    private Map map;
-    private List list;
-
-    private Forte forte;
-
-
-    private class Forte {
-        private String name;
-        private Integer time;
-
-        public String getName() {
-            return name;
-        }
-
-        public void setName(String name) {
-            this.name = name;
-        }
-
-        public Integer getTime() {
-            return time;
-        }
-
-        public void setTime(Integer time) {
-            this.time = time;
-        }
-    }
-
-    public String getLastName() {
-        return lastName;
-    }
-
-    public void setLastName(String lastName) {
-        this.lastName = lastName;
-    }
-
-    public Integer getAge() {
-        return age;
-    }
-
-    public void setAge(Integer age) {
-        this.age = age;
-    }
-
-    public Double getSalary() {
-        return salary;
-    }
-
-    public void setSalary(Double salary) {
-        this.salary = salary;
-    }
-
-    public Boolean getBoss() {
-        return boss;
-    }
-
-    public void setBoss(Boolean boss) {
-        this.boss = boss;
-    }
-
-    public Date getBirthday() {
-        return birthday;
-    }
-
-    public void setBirthday(Date birthday) {
-        this.birthday = birthday;
-    }
-
-    public Map getMap() {
-        return map;
-    }
-
-    public void setMap(Map map) {
-        this.map = map;
-    }
-
-    public List getList() {
-        return list;
-    }
-
-    public void setList(List list) {
-        this.list = list;
-    }
-
-    public Forte getForte() {
-        return forte;
-    }
-
-    public void setForte(Forte forte) {
-        this.forte = forte;
-    }
-
-    @Override
-    public String toString() {
-        return "Emp{" +
-                "lastName='" + lastName + '\'' +
-                ", age=" + age +
-                ", salary=" + salary +
-                ", boss=" + boss +
-                ", birthday=" + birthday +
-                ", map=" + map +
-                ", list=" + list +
-                ", forte=" + forte +
-                '}';
-    }
-}
