--- src/main/java/com/rockheart/po/User.java
@@ -1,40 +1,0 @@
-package com.rockheart.po;
-
-public class User {
-
-    private String username;
-    private int gender;
-
-    public User() {
-    }
-
-    public User(String username, int gender) {
-        this.username = username;
-        this.gender = gender;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
-
-    public int getGender() {
-        return gender;
-    }
-
-    public void setGender(int gender) {
-        this.gender = gender;
-    }
-
-    @Override
-    public String toString() {
-        return "User{" +
-                "username='" + username + '\'' +
-                ", gender=" + gender +
-                '}';
-    }
-
-}
