--- src/main/java/com/rockheart/service/EmpService.java
@@ -1,10 +1,0 @@
-package com.rockheart.service;
-
-public class EmpService{
-
-    public void add() {
-        //todo
-    }
-
-
-}
