--- src/main/resources/config/application.properties
@@ -1,12 +1,0 @@
-#emp.last-name=test emp
-#emp.age=30
-#emp.birthday=1989/9/12
-#emp.boss=false
-#emp.salary=23000
-#emp.map.key1=value1
-#emp.map.key2=value2
-#emp.list=one, two, three
-#emp.forte.name=python
-#emp.forte.time=3
-
-spring.profiles.active=dev
