--- src/main/resources/config/emp.yml
@@ -1,16 +1,0 @@
-emp:
-  lastName: zhaolei152624@163.com
-  age: 28
-  salary: 20000
-  boss: true
-  birthday: 1991/10/10
-  map:
-    key1: value1
-    key2: value2
-  list:
-    - one
-    - two
-    - three
-  forte:
-    name: java
-    time: 8
